% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgov_functions.R
\name{ctgov_ae_tables_multi_arm}
\alias{ctgov_ae_tables_multi_arm}
\title{Combine CT.GOV Adverse Events Tables from Multiple Arms}
\usage{
ctgov_ae_tables_multi_arm(ae_arm_data_list)
}
\arguments{
\item{ae_arm_data_list}{A list of all the AE tables you created by arm using
the single arm function. Example: ae_arm_data_list <- list(arm1, arm2, arm3)}
}
\value{
A data frame/tibble object with combined adverse events data from all arms
}
\description{
Combines adverse events tables from multiple treatment arms into a single
CT.GOV formatted table.
}
\note{
Ensure that the order of the list is the order of the arms on the template
}
