% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgov_functions.R
\name{ctgov_ae_tables_single_arm}
\alias{ctgov_ae_tables_single_arm}
\title{Create CT.GOV Adverse Events Table for Single Arm}
\usage{
ctgov_ae_tables_single_arm(
  processed_ae_data,
  ptid_ae_evaluable,
  trt_name = "drug",
  sae_or_other = "Other"
)
}
\arguments{
\item{processed_ae_data}{Your final analysis dataset for the reporting you wish to complete.
Must have variables: "ae_ctcae_cat", "ae_ctcae_term_value", and "ptid"}

\item{ptid_ae_evaluable}{A vector of patient IDs who are evaluable for the toxicity/safety
endpoint, also defined as "participants included in the assessment of adverse events
(that is, the denominator for calculating frequency of adverse events)"}

\item{trt_name}{The name/identifier of the treatment for this arm, names column.
Default is "drug"}

\item{sae_or_other}{Either "Serious" or "Other", for what type of CT.GOV AE report you want.
Default is "Other"}
}
\value{
A data frame/tibble object formatted as a CT.GOV adverse events table
}
\description{
Generates a CT.GOV formatted adverse events table for a single treatment arm,
either for serious adverse events (SAEs) or other adverse events (non-SAEs).
}
\note{
"Other" simply means not serious adverse event (non-SAE)

If end goal is to use export function, ensure that trt_name matches the template

You can filter your datasets within this function to input into the multi-arm
  function that will combine them together
}
