% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgov_functions.R
\name{exporting_xlsx_template}
\alias{exporting_xlsx_template}
\title{Export CT.GOV Adverse Events Table to Excel Template}
\usage{
exporting_xlsx_template(
  template_dir,
  template_file,
  ctgov_ae_table,
  dir_out,
  study_number,
  sae_or_other
)
}
\arguments{
\item{template_dir}{The file folder of the template given for this study}

\item{template_file}{The file name of the template given for this study}

\item{ctgov_ae_table}{Data frame object that is formatted as a CT.GOV AE table
(preferably created from the functions above)}

\item{dir_out}{The directory of the folder you wish the templates to be exported to}

\item{study_number}{A string of your study number}

\item{sae_or_other}{Either "serious" or "other" for what type of CT.GOV AE table}
}
\value{
Saves an Excel file to the specified directory. Returns a message if
  column names do not match template
}
\description{
Exports a CT.GOV formatted adverse events table to an Excel template file.
}
\note{
"Other" simply means not serious adverse event (non-SAE)
}
