% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_names.R
\name{get_cdisc_labels}
\alias{get_cdisc_labels}
\title{Load CDISC Labels for Specified Dataset}
\usage{
get_cdisc_labels(
  dataset_code,
  metadata_folder,
  dict_file = "2022-04-14 TB-PACTS_Data_Dictionary.xlsx",
  sheet = "Variable Descriptions",
  cols = 1:3
)
}
\arguments{
\item{dataset_code}{Character string. Lowercase dataset abbreviation (e.g., "dm", "ae", "lb")}

\item{metadata_folder}{Character string. Path to folder containing the data dictionary}

\item{dict_file}{Character string. Name of the data dictionary Excel file.
Default is "2022-04-14 TB-PACTS_Data_Dictionary.xlsx"}

\item{sheet}{Character string. Name of the sheet containing variable descriptions.
Default is "Variable Descriptions"}

\item{cols}{Numeric vector. Columns to read. Default is 1:3 (Variable, Label, Type)}
}
\value{
A data frame with CDISC variable labels, or NULL if dataset code not found
}
\description{
Reads CDISC variable labels from the TB-PACTS Data Dictionary for a specified dataset.
}
\note{
Available dataset codes include: dm, ae, be, bs, ce, cm, co, da, dd, di, ds, dv,
  ec, eg, er, ex, fa, ho, ie, lb, mb, mh, mi, mo, ms, oe, pc, pe, pf, pp, pr, qs, re,
  rp, sc, se, sr, ss, su, sv, vs, xo, relrec, ta, td, te, ti, ts, tv, apdm, apmh
}
