% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgov_functions.R
\name{pre_processing_ae_data}
\alias{pre_processing_ae_data}
\title{Pre-process Adverse Event Data}
\usage{
pre_processing_ae_data(ae_data, cutoff_dt = dt_consent)
}
\arguments{
\item{ae_data}{Your final analysis dataset for the reporting you wish to complete.
Must have variables: "ae_st_date", "ae_ctcae_cat", "ae_ctcae_term_value",
"dt_consent", and "ptid"}

\item{cutoff_dt}{A variable holding cutoff dates for when AEs should have started
to be included in reporting (defaulted to dt_consent)}
}
\value{
A data frame with filtered and renamed adverse event data
}
\description{
This function filters out baseline data using an inputted cutoff date
and ae_st_date, as well as renames two variables. It defaults to using
dt_consent if you do not specify a cutoff date.
}
\note{
cutoff_dt and ae_st_date must be formatted as dates before input

If you do not wish to use this method for filtering out baseline,
  please manually filter out baseline and rename variables
}
