% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_analysis_helpers.R
\name{standardize_flag}
\alias{standardize_flag}
\title{Standardize population flags to Yes/No format}
\usage{
standardize_flag(x)
}
\arguments{
\item{x}{Vector of flag values}
}
\value{
Character vector with "Yes", "No", or NA
}
\description{
Converts various flag coding schemes to standardized "Yes"/"No" values.
}
