% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_analysis_helpers.R
\name{standardize_sae}
\alias{standardize_sae}
\title{Standardize SAE indicator to Yes/No format}
\usage{
standardize_sae(x)
}
\arguments{
\item{x}{Vector of SAE values}
}
\value{
Character vector with "Yes", "No", or NA
}
\description{
Converts various SAE coding schemes to standardized "Yes"/"No" values.
Handles: Y/N, 1/0, Yes/No, TRUE/FALSE
}
\examples{
standardize_sae(c("Y", "N", "Y", NA))
}
