% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_analysis_helpers.R
\name{summarize_by_arm}
\alias{summarize_by_arm}
\title{Summarize AE data by treatment arm}
\usage{
summarize_by_arm(df)
}
\arguments{
\item{df}{Analysis dataset with arm_code, ptid, ae_sae columns}
}
\value{
Data frame with summary statistics per arm
}
\description{
Summarize AE data by treatment arm
}
\examples{
summarize_by_arm(ae_analysis)
}
